### 2021-10-26 #3851 @whedon
[PRE REVIEW]: PopMedNet: A scalable and extensible open-source informatics platform designed to facilitate the implementation and operation of distributed health data networks
**Submitting author:** @ddeehere (<a href="http://orcid.org/0000-0001-7021-6227">Daniel Dee</a>)
**Repository:** <a href="https://github.com/PopMedNet-Team/popmednet.git" target ="_blank">https://github.com/PopMedNet-Team/popmednet</a>
**Version:** 2021.4
