### 2021-11-02 #3851 @danielskatz
👋 @ddeehere - I think you missed that fact that the invitation to edit was to @csoneson. JOSS reviews include a number of roles, including
* an Assistant Editor-in-Chief (AEiC), currently me
* an editor, to be found by the AEiC, not currently filled
* at least two reviewers, to be found by the editor, not currently filled
* the author, you
