### 2021-11-03 #3851 @danielskatz
@ddeehere - I'll be the editor.  The next step is to find at least 2 reviewers.  If you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).
