### 2021-11-04 #3851 @danielskatz
external are fine - the list is just one resource if useful
(of course, please keep in mind the [JOSS conflict of interest guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy))
