### 2021-11-08 #3851 @danielskatz
👋 @martinmodrak & @borishejblum - would one of both of you be willing to review this JOSS submission?
