### 2021-11-08 #3851 @ddeehere
> @ddeehere - any further suggestions?

For external reviewers, would a developer who works for the same company (but not directly involved in the PopMedNet development) be acceptable? How about a subcontractor for Harvard Pilgrim Health Care Institute (who's the primary mover of PopMedNet) be acceptable? I am thinking in terms of whether that will violate conflict of interest guidelines. Neither of them are actively involved in PopMedNet.
