### 2021-11-08 #3851 @danielskatz
someone from the same company would not make sense, but someone who is a user would
