### 2021-11-08 #3851 @martinmodrak
Unfortunately, this is both at the very edge of my knowledge (I did some WinForms a long time ago, but have never used ASP.NET, I also now only bare minimum on security) and quite a large project - I won't be able to set aside time to review it responsibly at the moment.
