### 2021-11-15 #3851 @danielskatz
👋 @scossin & @vianneyJouhet - just checking with you both again re https://github.com/openjournals/joss-reviews/issues/3851#issuecomment-963542710
