### 2021-12-01 #3851 @ddeehere
@danielskatz I have reached out and waiting to hear from potential reviewers from outside JOSS volunteers.
