+++ Tests/PopMedNet.UITests/LoginPageTests_NoLogin.cs
@@ -1,0 +1,208 @@
+using NUnit.Framework;
+using Microsoft.Playwright;
+using System.Threading.Tasks;
+using System;
+using PopMedNet.UITests.Models;
+using System.Configuration;
+
+namespace PopMedNet.UITests
+{
+    public class LoginPageTests_NoLogin
+    {
+        IPlaywright playwright;
+        IBrowser browser;
+        IBrowserContext context;
+        IPage singlePage;
+
+        string testUrl;
+
+        public async Task<IPage> GetPage()
+        {
+            playwright = await Playwright.CreateAsync();
+            browser = await playwright.Chromium.LaunchAsync(new BrowserTypeLaunchOptions
+            {
+                //Headless = false,
+                //SlowMo = 50,
+            });
+            context = await browser.NewContextAsync();
+            return await context.NewPageAsync();
+        }
+
+        [OneTimeSetUp]
+        public void Setup()
+        {
+
+            testUrl = ConfigurationManager.AppSettings["baseUrl"];
+
+            singlePage = GetPage().Result;
+
+        }
+
+        [Test]
+        public async Task LaunchPMN()
+        {
+            // Given - nothing extra here
+
+            // When
+            await singlePage.GotoAsync(testUrl);
+
+            // Then
+            Assert.That(singlePage.Url, Does.StartWith("http://localhost:60338/"));
+        }
+
+
+        [Test]
+        public async Task Login_NoUsernameEntered_UserNameValidationMessageDisplays()
+        {
+            // Given
+            var loginPage = new LoginPage(singlePage);
+            await loginPage.Goto();
+
+            // When            
+            await loginPage.LoginAs("", "FakePass");
+
+            // Then
+            var userNameValText = await singlePage.WaitForSelectorAsync("#username-error");
+            Assert.That(userNameValText.InnerTextAsync, Is.EqualTo("username is required"));
+        }
+
+        [Test]
+        public async Task Login_NoUsernameOrPasswordEntered_UserNameAndPasswordValidationMessagesDisplay()
+        {
+            // Given
+            await singlePage.GotoAsync(testUrl);
+
+            // When
+
+            await singlePage.ClickAsync("#btnLogin");
+
+            // Then
+            var userNameValText = await singlePage.WaitForSelectorAsync("#username-error");
+            var passwordValText = await singlePage.WaitForSelectorAsync("#password-error");
+
+            Assert.That(userNameValText.InnerTextAsync, Is.EqualTo("username is required"),
+                "Username validation message did not display as expected.");
+            Assert.That(passwordValText.InnerTextAsync, Is.EqualTo("password is required"),
+                "Password validation message did not display as expected.");
+        }
+
+        [Test]
+        public async Task ContactUs_NavigatesToPMNSupportDesk()
+        {
+            // Given
+            await singlePage.GotoAsync(testUrl);
+
+            // When
+            var newPage = await context.RunAndWaitForPageAsync(async () =>
+            {
+                await singlePage.ClickAsync("text=Contact Us");
+            });
+
+            // Then 
+            await newPage.WaitForLoadStateAsync();
+            var title = await newPage.TitleAsync();
+            Assert.That(title, Is.EqualTo("PopMedNet Support - Jira Service Management"));
+        }
+
+        [Test]
+        public async Task TermsAndConditions_DisplaysTandCDialog()
+        {
+            // Given
+            await singlePage.GotoAsync(testUrl);
+
+            // When
+            await singlePage.ClickAsync("#Content :text(\"Terms and Conditions\")");
+
+            // Then 
[…]
