+++ Lpp.Dns.Api/Users/UserDeactivationJob.cs
@@ -34,7 +34,7 @@ namespace Lpp.Dns.Api.Users
         /// </summary>
         /// <remarks>The Hangfire job is set to attempt only once, and on fail only mark the job as failed.</remarks>
         /// <returns></returns>
-        [Hangfire.AutomaticRetry(Attempts = 1, OnAttemptsExceeded = Hangfire.AttemptsExceededAction.Fail)]
+        [Hangfire.AutomaticRetry(Attempts = 3, OnAttemptsExceeded = Hangfire.AttemptsExceededAction.Fail)]
         public async Task DeactivateStaleUsers()
         {
             int deactivateAfterDays = int.Parse(WebConfigurationManager.AppSettings["Users.DeactivateAfterNumberOfDays"]);
@@ -46,7 +46,9 @@ namespace Lpp.Dns.Api.Users
                                     select
                                     new
                                     {
-                                        User = u,
+                                        UserID = u.ID,
+                                        UserName = u.UserName,
+                                        OrganizationID = u.OrganizationID,
                                         Organization = u.Organization.Name,
                                         OrganizationAcronym = u.Organization.Acronym,
                                         LastSuccessfullAuthentication = _db.LogsUserAuthentication.Where(l => l.Success && l.UserID == u.ID).OrderByDescending(l => l.TimeStamp).FirstOrDefault()
@@ -59,36 +61,39 @@ namespace Lpp.Dns.Api.Users
             {
                 foreach (var su in staleUsers)
                 {
-                    var user = su.User;
-
-                    user.Active = false;
-                    user.DeactivatedOn = DateTime.UtcNow;
-                    user.DeactivationReason = $"User was deactivated via system service due to not having successfully authenticated within {deactivateAfterDays} days.";
-
-                    var validationErrors = _db.GetValidationErrors();
-                    if (validationErrors.Any())
+                    using (var database = new DataContext())
                     {
-                        //log validation error to email
-                        var errors = validationErrors.SelectMany(err => err.ValidationErrors.Select(ex => ex.ErrorMessage)).ToArray();
-                        validationErrorUsers.Add(user.ID, errors);
+                        var user = await database.Users.FindAsync(su.UserID);
+                        if(user == null)
+                        {
+                            continue;
+                        }
 
-                        continue;
+                        user.Active = false;
+                        user.DeactivatedOn = DateTime.UtcNow;
+                        user.DeactivationReason = $"User was deactivated via system service due to not having successfully authenticated within {deactivateAfterDays} days.";
+
+                        var validationErrors = database.GetValidationErrors();
+                        if (validationErrors.Any())
+                        {
+                            //log validation error to email
+                            var errors = validationErrors.SelectMany(err => err.ValidationErrors.Select(ex => ex.ErrorMessage)).ToArray();
+                            validationErrorUsers.Add(user.ID, errors);
+
+                            continue;
+                        }
+
+                        database.LogsUserChange.Add(new UserChangeLog
+                        {
+                            UserID = Guid.Empty,
+                            UserChangedID = user.ID,
+                            Reason = EntityState.Modified,
+                            Description = $"User '{ su.OrganizationAcronym }\\{ user.UserName }' was deactivated via system service due to not having successfully authenticated within { deactivateAfterDays } days."
+                        });
+
+                        await database.SaveChangesAsync();
                     }
-
[…]
