### 2021-12-23 #3851 @danielskatz
@ddeehere - every published paper from JOSS (https://joss.theoj.org/papers/published) has a link on the right side for Paper review.  In addition, https://joss.readthedocs.io/en/latest/reviewer_guidelines.html provides general guidelines.
