### 2022-01-13 #4062 @whedon
[REVIEW]: PopMedNet: A scalable and extensible open-source informatics platform designed to facilitate the implementation and operation of distributed health data networks
**Submitting author:** @ddeehere (<a href="http://orcid.org/0000-0001-7021-6227">Daniel Dee</a>)
**Repository:** https://github.com/PopMedNet-Team/popmednet
**Branch with paper.md** (empty if default branch):
**Version:** PMN-2021.6-JOSS

- [x] #4119
