### 2022-01-14 PopMedNet-Team/popmednet#31 @jhancock4d
LICENSE.txt should be renamed as just LICENSE
Naming is a linux thing. No extension.

Right now the file links externally to another license. Best practice is that you have the link to the source but also include the text.

openjournals/joss-reviews#4062
