+++ README.md
@@ -2,9 +2,39 @@
 
 PopMedNet&trade; is an open-source application used to facilitate multi-site health data networks. It uses a distributed network design that enables data holders to retain full control of their data. Investigators send questions to data holders for review and response. PopMedNet eliminates the need for assembling patient records in a centralized repository, thus preserving patient privacy and confidentiality.
 
-# To set up an instance of PopMedNet: 
+# Statement of need
 
-* Instructions are located here: https://popmednet.atlassian.net/wiki/spaces/DOC/pages/938409985/Building+Open+Source+PopMedNet
-* The starter databases are located here: https://popmednet.atlassian.net/wiki/spaces/DOC/pages/124685521/Starter+Databases
+Distributed health network plays an important role in supporting a number of population health activities. 
+
+The goal of PopMedNet&trade; is to facilitate distributed analysis of electronic health data in order to support medical product safety surveillance, disease surveilance, comparative effectiveness, quality, medical resource use, cost-effectiveness, and related studies. It does this by enabling the creation of health data networks that allows the data owners to keep and control uses of their data, while at the same time allows customized access and governance for each network.
+
+The concerns of the network stakeholders are addressed in the design and governance models of the PopMedNet&trade; software platform. 
+
+PopMedNet&trade; features include distributed querying, customizable workflows, and auditing and search capabilities. Its flexible role-based access control system enables the enforcement of varying governance policies.
+
+# To set up an instance of PopMedNet&trade;
+
+* Installation instructions are located at [Building Open Source PopMedNet&trade;](https://popmednet.atlassian.net/wiki/spaces/DOC/pages/938409985/Building+Open+Source+PopMedNet).
+* The starter databases are located at [Starter Databases](https://popmednet.atlassian.net/wiki/spaces/DOC/pages/124685521/Starter+Databases).
     * Please note the that PMN-Starter-2020.5.bak file is the most up to date backup.   
-* Please reach out to us at https://popmednet.atlassian.net/servicedesk/customer/portal/1/group/10 to report any problems.
+
+# Reporting issues
+Please reach out to us at [PopMedNet&trade; Service Desk](https://popmednet.atlassian.net/servicedesk/customer/portal/1/group/10) to report any problems and to seek software assistance.
+
+# Contribution
+We are currently still working on a process to accept external contribution. We will provide additional information here once that process is in place.
+
+In the meantime, you may submit contribution request to the support link above.
+
+# Usage documentation
+
+For end user of PopMedNet&trade;, the usage documentation is [PopMedNet&trade; User's Guide](https://popmednet.atlassian.net/wiki/spaces/DOC/pages/8880241/PopMedNet+User+s+Guide).
+
+# Software Paper
+An overview of PopMedNet&trade; can be found in the accompanying [software paper](paper.md).
+
+# Core functionality
+For a description of the core functionality, see the  section "How does PopMedNet&trade; work?" in the accompanying [software paper](paper.md).
+
+
+
