### 2022-01-25 PopMedNet-Team/popmednet#31 @ddeehere
LICENSE.txt renamed to LICENSE. Apache 2.0 license text included in LICENSE.
