### 2022-02-01 #4062 @danielskatz
@jhancock4d - thanks!

Re testing, see https://joss.readthedocs.io/en/latest/review_criteria.html#tests
Does this meet the "OK" there?
