### 2022-02-07 #4062 @danielskatz
@jhancock4d - can you say a bit more? Is this something where you are unsure about what is needed, or where you know what's needed and are unsure if it's being met?  Is https://joss.readthedocs.io/en/latest/review_criteria.html#tests helpful?
