### 2022-02-08 PopMedNet-Team/popmednet#35 @lrasmus
Clarifications in paper
[Paper](https://github.com/PopMedNet-Team/popmednet/blob/master/paper.md) is generally well-written and clear.

As part of openjournals/joss-reviews#4062 JOSS guidelines ask for "_State of the field: Do the authors describe how this software compares to other commonly-used packages?_"  I do not see comparison to other used packages and software (i2b2/SHRINE, OMOP/OHDSI stack, Leaf, etc.)  Could you please add in this background and comparison?

Minor suggestions:
* "Provide protection of PHI" - could you please define PHI (with brief explanation/references, such as HIPAA Privacy Regs)? (line 52)
* References for FDA Sentinel and PCORnet (lines 64-65)
