### 2022-02-09 PopMedNet-Team/popmednet#36 @lrasmus
Clarifications/updates in installation instructions (Required Software)
As part of https://github.com/openjournals/joss-reviews/issues/4062, I am reviewing the [installation instructions](https://popmednet.atlassian.net/wiki/spaces/DOC/pages/938409985/Building+Open+Source+PopMedNet) and had a few questions/requests for clarifications in the Required Software section:

1. "Visual Studio Professional 2019" - is Professional required?  Or will Community work (I am using Community, will report back if any hard stops, but seems to be working so far)
2. ".NET Framework 4.7.2 Dev Pack" - I received a warning when loading `Lpp.Adapters.sln` that it is targeting `v4.8`.  Should this reference 4.8 instead of 4.7.2?  When I install 4.8 I don't get warnings.
3. "MVC4 from https://www.microsoft.com/en-us/download/details.aspx?id=30683." - the [paper](https://github.com/PopMedNet-Team/popmednet/blob/master/paper.md) says it uses MVC5, not 4
