### 2022-02-09 PopMedNet-Team/popmednet#37 @lrasmus
Clarification in installation instructions [Building PopMedNet]
As part of https://github.com/openjournals/joss-reviews/issues/4062, I am reviewing the installation instructions and have one request under the "Building PopMedNet" section:
* For the final step (Login to PMN Portal as SystemAdministrator), could you please provide or add a link to where we can find the password for the SystemAdministrator (I was able to figure it out, but would like to have this easy to find for users).
