### 2022-02-09 PopMedNet-Team/popmednet#38 @lrasmus
Failing unit tests (setup steps required?)
As part of https://github.com/openjournals/joss-reviews/issues/4062, I am running through unit tests and had multiple failures.  The error is fairly consistent and descriptive:

```
    Unable to create instance of class Lpp.Dns.DataMart.Model.QueryComposer.Tests.SummaryQueryTests. Error: System.Configuration.ConfigurationErrorsException: Unable to open configSource file 'ConnectionStrings.config'. (C:\Development\popmednet\Lpp.Adapters\Lpp.Dns.DataMart.Model.QueryComposer.Tests\bin\Lpp.Dns.DataMart.Model.QueryComposer.Tests.dll.config line 97).
```

I wanted to confirm (I did not see documentation on the README page) what setup steps are needed to run the unit tests.  Are these more integration tests - requiring a full database setup, and setup of a configuration file?
