### 2022-02-10 PopMedNet-Team/popmednet#36 @ddeehere
Visual Studio Community is sufficient. It is not necessary to use the Professional edition.
We are now targeting .NET 4.8. 4.7.2 is no longer supported.
We are now using MVC5, though MVC4 should still work.
I will make these corrections in the documentation.
