### 2022-02-10 PopMedNet-Team/popmednet#38 @ddeehere
They do need a database setup with sample CDM (common data model) data. The unit tests do not make any modification to the database. You will need to be provided with this test data. We are planning to submit that to IEEE Data Port. In the meantime, I will figure out how to provide it to you.

The ConnectionString.config will have to be copied from the template and modified accordingly.
