### 2022-02-11 11fb836
JOSS review ticket #35 Define PHI, add ref to Sentinel & PCORNET
