### 2022-02-14 #4062 @lrasmus
@jhancock4d - I'm working with the authors via https://github.com/PopMedNet-Team/popmednet/issues/38 on setup instructions for running the unit tests.  Curious if you had a CDM database you used for the tests, if you used SynPuf or something else?  Am going to try to get this going on my own, and I know the authors are working on instructions.
