### 2022-02-14 #4062 @jhancock4d
> @jhancock4d - I'm working with the authors via [PopMedNet-Team/popmednet#38](https://github.com/PopMedNet-Team/popmednet/issues/38) on setup instructions for running the unit tests. Curious if you had a CDM database you used for the tests, if you used SynPuf or something else? Am going to try to get this going on my own, and I know the authors are working on instructions.

I just ran the migrations which got me a blank database and worked from there to setup everything by hand.  (I have an SQL DB server laying around)

The setup of the database probably needs better documentation. However, my main issue with testing is that there aren't very MANY tests, not that they don't work (they seem to)
