### 2022-02-15 PopMedNet-Team/popmednet#38 @lrasmus
Thank you @ddeehere!  Do the tests expect your specific dataset, or is it possible to incorporate - for example - an OMOP CDM populated with SynPuf data?  If there are general instructions (apologies if I missed these) on how to configure and run the tests, I'm happy to keep proceeding.
