### 2022-02-15 PopMedNet-Team/popmednet#38 @ddeehere
@lrasmus I added a section on Regression Tests in the README.md, which links to the Confluence page with an explanation of the framework and the steps to run it. The way the tests is collected is this:

* New features are manually tested.
* Once verified, the query and the results (request and response) are collected as JSON files and added to a folder called Resources, and a new test method is added as well. Normally, a test method simply calls a common method in the superclass.
* When run, the request file is loaded and run through the engine. The outcome is then compared against the response file.
* If there is a discrepancy, then we know that either the data has changed or a bug has been introduced.

Currently, we do not support OMOP.

NOTE: Still working on providing the CDM.
