### 2022-02-15 PopMedNet-Team/popmednet#36 @ddeehere
@lrasmus Documentation updated. MVC5 is integrated into the solution and does not have to be downloaded separately anymore.
