### 2022-02-15 PopMedNet-Team/popmednet#39 @lrasmus
"State of the Field" in paper
As part of https://github.com/openjournals/joss-reviews/issues/4062 JOSS guidelines ask for "_State of the field: Do the authors describe how this software compares to other commonly-used packages?_"

I do not see comparison to other used packages and software (i2b2/SHRINE, OMOP/OHDSI stack, Leaf, etc.) Could you please add this to the paper?
