### 2022-02-16 #4062 @jhancock4d
@irasmus The integration tests are actually a feature not a bug in this case I would say because unit tests on a project that deals with data that would use mocks wouldn't be useful.

Code coverage is relatively lowish from what I can tell (I can't get coverlet to spit out data on it).  I think it's fine for this purpose, but in general it would be nice to see 80%+ code coverage in the future with coverlet commands that would do code coverage reporting along with the testing.
