### 2022-02-21 PopMedNet-Team/popmednet#39 @ddeehere
@lrasmus Just want to let you know that I am still working on this. I do have some experience with I2B2 and Shrine, but it may be dated. So I need to brush up on their recent development.
