### 2022-02-21 #4062 @danielskatz
👋 @ddeehere - I think we're just waiting for you to address https://github.com/PopMedNet-Team/popmednet/issues/39 at this point, and I see there that you are working on it. Please let us know when there's something for the reviewers to check.
