### 2022-02-22 643d2d7
Update paper to include comparison with i2b2/SHRINE
