+++ paper.bib
@@ -32,6 +32,16 @@
     url = {https://institutionalrepository.aah.org/jpcrr/vol2/iss2/99/}
 }
 
+@article{McMurry:2013,
+    author = {{McMurry}, A.~J. and {Murphy}, S.~N. and {MacFadden}, D. and {Griffin}, W.} and {Simons}, W.~W. and {Orcechia}, J. and {Bickel}, J. and {Wattanasin}, N. and {Gilbert}, C. and {Trevvett}, P. and {Churchill}, S. and {Kohane}, I.~S.,
+    title = "{SHRINE: Enabling Nationally Scalable Multi-Site Disease Studies}",
+    journal = {PLOS ONE},
+    keywords = {distributed research network, metadata},
+    year = 2013,
+    month = mar,
+    url = {https://catalyst.harvard.edu/wp-content/uploads/2021/03/SHRINE-PLOS_ONE_paper.pdf}
+}
+
 @misc{PopMedNet, 
     title = {PopMedNet},
     url={https://www.popmednet.org/}, 
