### 2022-02-22 PopMedNet-Team/popmednet#39 @ddeehere
@lrasmus I added a section "Comparison to i2b2 and SHRINE" for "State of the Field" in paper.md.
