### 2022-02-22 #4062 @ddeehere
@danielskatz @lrasmus I added a section "Comparison with i2b2 and SHRINE" for "State of the Field" in paper.md.
