### 2022-02-22 f58cedf
fixes to bib file
