+++ paper.bib
@@ -1,6 +1,6 @@
 @article{Davies:2016,
-  	Author = {{Davies}, M. and {Erickson}, K. and {Wyner}, Z. and {Malenfant}, J.~M. and {Rosen}, R. and {Brown}, J.},
-  	Title = {Software-Enabled Distributed Network Governance: The PopMedNet Experience},
+  	Author = {Davies}, M. and {Erickson}, K. and {Wyner}, Z. and {Malenfant}, J.~M. and {Rosen}, R. and {Brown}, J.},
+  	Title = {Software-Enabled Distributed Network Governance: The {PopMedNet} Experience},
   	Journal = {The Journal of Electronic Health Data and Methods},
   	Keywords = {distributed network, governance},
   	Month = mar,
@@ -11,7 +11,7 @@
 
 @article{Brown:2010,
     author = {{Brown}, J. and {Holmes} and J.~H. and {Shah}, K. and {Hall}, K. and {Lazarus}, R. and {Platt}, R.},
-    title = "{Distributed Health Data Networks: A Practical and Preferred Approach to Multi-Institutional Evaluations of Comparative Effectiveness, Safety, and Quality of Care}",
+    title = {Distributed Health Data Networks: A Practical and Preferred Approach to Multi-Institutional Evaluations of Comparative Effectiveness, Safety, and Quality of Care},
     journal = {Medical Care},
     keywords = {health care, multi-institutional},
     year = 2010,
@@ -23,7 +23,7 @@
 
 @article{Malenfant:2015,
     author = {{Malenfant}, J. and {Davies}, M. and {Swan}, B. and {Brown}, J.~S.},
-    title = "{PopMedNet: Collecting and Using Metadata in Distributed Research Networks}",
+    title = {{PopMedNet}: Collecting and Using Metadata in Distributed Research Networks},
     journal = {Journal of Patient-Centered Research and Reviews},
     keywords = {distributed research network, metadata},
     year = 2015,
@@ -33,8 +33,8 @@
 }
 
 @article{McMurry:2013,
-    author = {{McMurry}, A.~J. and {Murphy}, S.~N. and {MacFadden}, D. and {Griffin}, W.} and {Simons}, W.~W. and {Orcechia}, J. and {Bickel}, J. and {Wattanasin}, N. and {Gilbert}, C. and {Trevvett}, P. and {Churchill}, S. and {Kohane}, I.~S.,
-    title = "{SHRINE: Enabling Nationally Scalable Multi-Site Disease Studies}",
+    author = {{McMurry}, A.~J. and {Murphy}, S.~N. and {MacFadden}, D. and {Griffin}, W.} and {Simons}, W.~W. and {Orcechia}, J. and {Bickel}, J. and {Wattanasin}, N. and {Gilbert}, C. and {Trevvett}, P. and {Churchill}, S. and {Kohane}, I.~S.},
+    title = {{SHRINE}: Enabling Nationally Scalable Multi-Site Disease Studies},
     journal = {PLOS ONE},
     keywords = {distributed research network, metadata},
     year = 2013,
@@ -43,7 +43,7 @@
 }
 
 @misc{PopMedNet, 
-    title = {PopMedNet},
+    title = {{PopMedNet}},
     url={https://www.popmednet.org/}, 
     journal={PopMedNet}, 
     publisher={Harvard Pilgrim Health Care Institute}
