+++ paper.bib
@@ -1,5 +1,5 @@
 @article{Davies:2016,
-  	Author = {Davies}, M. and {Erickson}, K. and {Wyner}, Z. and {Malenfant}, J.~M. and {Rosen}, R. and {Brown}, J.},
+  	Author = {{Davies}, M. and {Erickson}, K. and {Wyner}, Z. and {Malenfant}, J.~M. and {Rosen}, R. and {Brown}, J.},
   	Title = {Software-Enabled Distributed Network Governance: The {PopMedNet} Experience},
   	Journal = {The Journal of Electronic Health Data and Methods},
   	Keywords = {distributed network, governance},
