+++ paper.bib
@@ -33,7 +33,7 @@
 }
 
 @article{McMurry:2013,
-    author = {{McMurry}, A.~J. and {Murphy}, S.~N. and {MacFadden}, D. and {Griffin}, W.} and {Simons}, W.~W. and {Orcechia}, J. and {Bickel}, J. and {Wattanasin}, N. and {Gilbert}, C. and {Trevvett}, P. and {Churchill}, S. and {Kohane}, I.~S.},
+    author = {{McMurry}, A.~J. and {Murphy}, S.~N. and {MacFadden}, D. and {Griffin}, W. and {Simons}, W.~W. and {Orcechia}, J. and {Bickel}, J. and {Wattanasin}, N. and {Gilbert}, C. and {Trevvett}, P. and {Churchill}, S. and {Kohane}, I.~S.},
     title = {{SHRINE}: Enabling Nationally Scalable Multi-Site Disease Studies},
     journal = {PLOS ONE},
     keywords = {distributed research network, metadata},
