+++ paper.bib
@@ -39,6 +39,7 @@
     keywords = {distributed research network, metadata},
     year = 2013,
     month = mar,
+    doi = {10.1371/journal.pone.0055811},
     url = {https://catalyst.harvard.edu/wp-content/uploads/2021/03/SHRINE-PLOS_ONE_paper.pdf}
 }
 
