### 2022-02-23 #4062 @danielskatz
@ddeehere - proofreading the paper PDF, I found a couple of things that I suggest changing, as shown in https://github.com/PopMedNet-Team/popmednet/pull/44  Please merge this if you agree with these changes.
