### 2022-03-02 #4062 @ddeehere
> You might want to contact their support - this should be very fast when it works, which it always has in my experience

@danielskatz Just want to let you know that we are still looking into this. We have sent a support request to Zenodo.
