### 2022-03-03 #4062 @danielskatz
@ddeehere - Please check that the Zenodo item has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it - I should be removed). You may also add the authors' ORCID.

You can change metadata and then republish the zenodo item without getting or needing a new DOI, as long as you don't change the data itself (the zip file in this case)
