### 2022-03-03 #4062 @ddeehere
> See "How can I edit the metadata of a published record?" on https://help.zenodo.org

@danielskatz Thank you for all your help! Metadata is edited.
