### 2022-03-03 #4062 @xuanxu
It looks like both figures have the same id (fig:PopMedNet) in the paper.md file. Changing it in any of them should fix it.
