### 2022-03-03 0379219
Update figure label to make it unique
