### 2022-03-03 #4062 @danielskatz
@ddeehere - can you update this?  We don't need a new zenodo deposit, as we only require the software to be archived, not the final paper
