### 2022-03-04 #4062 @arfon
> @ddeehere - there are a couple of things I'm not happy about in the metadata and pdf, but these are issues on the JOSS side, not on what you submitted. Until these are fixed, I'm not going to to ahead with the final acceptance and publishing, but hopefully this won't take long to resolve

@danielskatz – these should be addressed now.
